refactor: extract memory, agent_hooks, telemetry, workflows, services into tools/
First batch of 5 modules extracted from fastmcp_server.py (52 tools total,
count unchanged). Each module exposes register(mcp); __init__.py composes
them via register_all(). Remaining 12 modules follow in later tasks.
