+++ iterm_mcpy/fastmcp_server.py
@@ -482,6 +482,11 @@ mcp = FastMCP(
     dependencies=["iterm2", "asyncio", "pydantic"]
 )
 
+# Register tool modules extracted from this file.
+# See iterm_mcpy/tools/__init__.py
+from iterm_mcpy.tools import register_all  # noqa: E402
+register_all(mcp)
+
 
 # ============================================================================
 # OAUTH METADATA ENDPOINTS (for HTTP transport compatibility)
@@ -554,51 +559,6 @@ from iterm_mcpy.helpers import (  # noqa: E402
     resolve_session,
     resolve_target_sessions,
 )
-
-
-async def _start_telemetry_server(port: int, duration: int = 300) -> str:
-    """Start a lightweight HTTP server that streams telemetry JSON."""
-
-    if _telemetry is None or _terminal is None:
-        raise RuntimeError("Telemetry not initialized")
-
-    global _telemetry_server_task
-
-    if _telemetry_server_task:
-        _telemetry_server_task.cancel()
-
-    async def handle(reader: asyncio.StreamReader, writer: asyncio.StreamWriter) -> None:
-        try:
-            await _terminal.get_sessions()
-            payload = _telemetry.dashboard_state(_terminal)
-            body = json.dumps(payload, indent=2)
-            response = (
-                "HTTP/1.1 200 OK\r\n"
-                "Content-Type: application/json\r\n"
-                f"Content-Length: {len(body.encode())}\r\n"
-                "Connection: close\r\n\r\n"
-                f"{body}"
-            )
-            writer.write(response.encode())
-            await writer.drain()
-        finally:
-            writer.close()
-            await writer.wait_closed()
-
-    async def serve() -> None:
-        server = await asyncio.start_server(handle, "0.0.0.0", port)
-        try:
-            async with server:
-                await asyncio.wait_for(server.serve_forever(), timeout=duration)
-        except asyncio.TimeoutError:
-            # Normal shutdown after duration
-            pass
-        finally:
-            server.close()
-            await server.wait_closed()
-
-    _telemetry_server_task = asyncio.create_task(serve())
-    return f"Telemetry web dashboard running at http://localhost:{port} for {duration}s"
 
 
 def ensure_model(model_cls, payload):
@@ -3642,267 +3602,9 @@ async def get_feedback_config(
 
 
 # ============================================================================
-# SERVICE MANAGEMENT TOOLS (consolidated)
+# SERVICE MANAGEMENT TOOLS
+# See iterm_mcpy/tools/services.py
 # ============================================================================
-
-@mcp.tool()
-async def manage_services(request: ManageServicesRequest, ctx: Context) -> str:
-    """Unified service management - consolidates 6 service tools into one.
-
-    Operations:
-    - list: List configured services (optional: repo_path, min_priority, include_status)
-    - start: Start a service (requires service_name; optional: repo_path)
-    - stop: Stop a service (requires service_name)
-    - add: Add new service (requires service_name, command; optional: priority, display_name, port, etc.)
-    - configure: Update service config (requires service_name; optional: priority, port, command, etc.)
-    - list_inactive: Get services that should be running but aren't (requires repo_path)
-
-    Args:
-        request: ManageServicesRequest with operation and relevant parameters
-
-    Returns:
-        JSON with operation result
-    """
-    service_manager = ctx.request_context.lifespan_context["service_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        op = request.operation
-
-        # LIST operation
-        if op == "list":
-            priority = None
-            if request.min_priority:
-                priority = ServicePriority.from_string(request.min_priority)
-
-            if request.repo_path:
-                services = service_manager.get_merged_services(request.repo_path, priority)
-            else:
-                global_registry = service_manager.load_global_config()
-                services = global_registry.services
-                if priority:
-                    priority_order = [
-                        ServicePriority.QUIET,
-                        ServicePriority.OPTIONAL,
-                        ServicePriority.PREFERRED,
-                        ServicePriority.REQUIRED
-                    ]
-                    min_idx = priority_order.index(priority)
-                    services = [
-                        s for s in services
-                        if priority_order.index(s.priority) >= min_idx
-                    ]
-
-            result = []
-            for service in services:
-                info = {
-                    "name": service.name,
-                    "display_name": service.effective_display_name,
-                    "priority": service.priority.value,
-                    "command": service.command,
-                    "port": service.port,
-                    "working_directory": service.working_directory,
-                }
-                if request.include_status:
-                    is_running = await service_manager.check_service_running(service)
-                    info["is_running"] = is_running
-                result.append(info)
-
-            return ManageServicesResponse(
-                operation=op,
-                success=True,
-                data={"services": result, "count": len(result), "repo_path": request.repo_path}
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # START operation
-        elif op == "start":
-            if not request.service_name:
-                raise ValueError("service_name is required for start operation")
-
-            if request.repo_path:
-                services = service_manager.get_merged_services(request.repo_path)
-            else:
-                global_registry = service_manager.load_global_config()
-                services = global_registry.services
-
-            service = None
-            for s in services:
-                if s.name == request.service_name:
-                    service = s
-                    break
-
-            if not service:
-                return ManageServicesResponse(
-                    operation=op,
-                    success=False,
-                    error=f"Service '{request.service_name}' not found",
-                    data={"available_services": [s.name for s in services]}
-                ).model_dump_json(indent=2, exclude_none=True)
-
-            state = await service_manager.start_service(service, repo_path=request.repo_path)
-
-            return ManageServicesResponse(
-                operation=op,
-                success=state.is_running,
-                data={
-                    "service": request.service_name,
-                    "started": state.is_running,
-                    "session_id": state.session_id,
-                    "error": state.error_message,
-                }
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # STOP operation
-        elif op == "stop":
-            if not request.service_name:
-                raise ValueError("service_name is required for stop operation")
-
-            success = await service_manager.stop_service(request.service_name)
-
-            return ManageServicesResponse(
-                operation=op,
-                success=success,
-                data={"service": request.service_name, "stopped": success}
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # ADD operation
-        elif op == "add":
-            if not request.service_name:
-                raise ValueError("service_name is required for add operation")
-            if not request.command:
-                raise ValueError("command is required for add operation")
-
-            service = ServiceConfig(
-                name=request.service_name,
-                display_name=request.display_name,
-                command=request.command,
-                priority=ServicePriority.from_string(request.priority or "optional"),
-                port=request.port,
-                working_directory=request.working_directory,
-                repo_patterns=request.repo_patterns or [],
-            )
-
-            if request.scope == "repo":
-                if not request.repo_path:
-                    return ManageServicesResponse(
-                        operation=op,
-                        success=False,
-                        error="repo_path required when scope is 'repo'"
-                    ).model_dump_json(indent=2, exclude_none=True)
-
-                registry = service_manager.load_repo_config(request.repo_path)
-                registry.services = [s for s in registry.services if s.name != request.service_name]
-                registry.services.append(service)
-                service_manager.save_repo_config(request.repo_path, registry)
-            else:
-                registry = service_manager.load_global_config()
-                registry.services = [s for s in registry.services if s.name != request.service_name]
-                registry.services.append(service)
-                service_manager.save_global_config(registry)
-
-            logger.info(f"Added service '{request.service_name}' to {request.scope} config")
-
-            return ManageServicesResponse(
-                operation=op,
-                success=True,
-                data={"service": request.service_name, "scope": request.scope, "added": True}
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # CONFIGURE operation
-        elif op == "configure":
-            if not request.service_name:
-                raise ValueError("service_name is required for configure operation")
-
-            if request.scope == "repo":
-                if not request.repo_path:
-                    return ManageServicesResponse(
-                        operation=op,
-                        success=False,
-                        error="repo_path required when scope is 'repo'"
-                    ).model_dump_json(indent=2, exclude_none=True)
-                registry = service_manager.load_repo_config(request.repo_path)
-            else:
-                registry = service_manager.load_global_config()
-
-            found = False
-            for i, service in enumerate(registry.services):
-                if service.name == request.service_name:
-                    found = True
-                    updates = {}
-                    if request.priority:
-                        updates["priority"] = ServicePriority.from_string(request.priority)
-                    if request.port is not None:
-                        updates["port"] = request.port
-                    if request.command:
-                        updates["command"] = request.command
-                    if request.working_directory:
-                        updates["working_directory"] = request.working_directory
-
-                    updated_data = service.model_dump()
-                    updated_data.update(updates)
-                    registry.services[i] = ServiceConfig.model_validate(updated_data)
-                    break
-
-            if not found:
-                return ManageServicesResponse(
-                    operation=op,
-                    success=False,
-                    error=f"Service '{request.service_name}' not found in {request.scope} config"
-                ).model_dump_json(indent=2, exclude_none=True)
-
-            if request.scope == "repo":
-                service_manager.save_repo_config(request.repo_path, registry)
-            else:
-                service_manager.save_global_config(registry)
-
-            logger.info(f"Updated service '{request.service_name}' in {request.scope} config")
-
-            return ManageServicesResponse(
-                operation=op,
-                success=True,
-                data={"service": request.service_name, "scope": request.scope, "updated": True}
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # LIST_INACTIVE operation
-        elif op == "list_inactive":
-            if not request.repo_path:
-                raise ValueError("repo_path is required for list_inactive operation")
-
-            priority = None
-            if request.min_priority:
-                priority = ServicePriority.from_string(request.min_priority)
-
-            inactive = await service_manager.get_inactive_services(request.repo_path, priority)
-
-            result = []
-            for service in inactive:
-                result.append({
-                    "name": service.name,
-                    "display_name": service.effective_display_name,
-                    "priority": service.priority.value,
-                    "command": service.command,
-                })
-
-            return ManageServicesResponse(
-                operation=op,
-                success=True,
-                data={"inactive_services": result, "count": len(result), "repo_path": request.repo_path}
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        else:
-            return ManageServicesResponse(
-                operation=op,
-                success=False,
-                error=f"Unknown operation: {op}"
-            ).model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error in manage_services ({request.operation}): {e}")
-        return ManageServicesResponse(
-            operation=request.operation,
-            success=False,
-            error=str(e)
-        ).model_dump_json(indent=2, exclude_none=True)
 
 
 # ============================================================================
@@ -4665,58 +4367,11 @@ async def get_sessions_by_role(
 
 
 # ============================================================================
-# TELEMETRY TOOL AND RESOURCE
+# TELEMETRY TOOL
+# See iterm_mcpy/tools/telemetry.py for start_telemetry_dashboard tool.
+# The telemetry_dashboard resource remains here (resources aren't part of
+# this extraction).
 # ============================================================================
-
-@mcp.tool()
-async def start_telemetry_dashboard(
-    ctx: Context,
-    port: int = 9999,
-    duration_seconds: int = 300,
-) -> str:
-    """Start a lightweight web server that streams telemetry JSON for external dashboards.
-
-    The dashboard provides:
-    - Real-time agent status cards with SSE updates
-    - Event stream showing notifications and activities
-    - Action buttons for focusing panes and sending commands via API calls
-    - Dark terminal theme matching iTerm2 aesthetic
-
-    Args:
-        port: Port to run the telemetry server on (default: 9999)
-        duration_seconds: How long to keep the server running (default: 300, 0 = indefinitely)
-    """
-    telemetry: TelemetryEmitter = ctx.request_context.lifespan_context["telemetry"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        message = await start_dashboard(
-            telemetry=telemetry,
-            terminal=_terminal,
-            notification_manager=_notification_manager,
-            port=port,
-            duration=duration_seconds,
-        )
-        logger.info(message)
-
-        # Include setup instructions
-        setup_msg = (
-            f"\n\nOpen the dashboard at: http://localhost:{port}\n\n"
-            f"The dashboard uses API calls for agent control:\n"
-            f"  - /api/focus?agent=<name> - Focus an agent's pane\n"
-            f"  - /api/send?agent=<name>&command=<cmd> - Send command to agent"
-        )
-
-        return json.dumps({
-            "status": "started",
-            "message": message,
-            "url": f"http://localhost:{port}",
-            "setup": setup_msg,
-        }, indent=2)
-    except Exception as e:
-        logger.error(f"Error starting telemetry server: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
 
 @mcp.resource("telemetry://dashboard")
 async def telemetry_dashboard() -> str:
@@ -4734,753 +4389,20 @@ async def telemetry_dashboard() -> str:
 
 # ============================================================================
 # WORKFLOW EVENT TOOLS
+# See iterm_mcpy/tools/workflows.py
 # ============================================================================
-
-# Priority level mapping
-PRIORITY_MAP = {
-    "low": EventPriority.LOW,
-    "normal": EventPriority.NORMAL,
-    "high": EventPriority.HIGH,
-    "critical": EventPriority.CRITICAL,
-}
-
-
-@mcp.tool()
-async def trigger_workflow_event(
-    ctx: Context,
-    event_name: str,
-    payload: Optional[Dict[str, Any]] = None,
-    source: Optional[str] = None,
-    priority: str = "normal",
-    metadata: Optional[Dict[str, Any]] = None,
-    immediate: bool = False
-) -> str:
-    """Trigger a workflow event.
-
-    Events are processed by registered listeners (@listen decorators) and can
-    be routed dynamically using @router decorators.
-
-    Args:
-        event_name: Name of the event to trigger
-        payload: Event payload data (will be passed to listeners)
-        source: Source of the event (agent/flow name)
-        priority: Event priority: low, normal, high, critical
-        metadata: Additional event metadata
-        immediate: If True, process synchronously instead of queueing
-
-    Returns:
-        JSON response with event info and processing result
-    """
-    event_bus: EventBus = ctx.request_context.lifespan_context["event_bus"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Map priority string to enum
-        priority_enum = PRIORITY_MAP.get(priority.lower(), EventPriority.NORMAL)
-
-        # Trigger the event
-        result = await event_bus.trigger(
-            event_name=event_name,
-            payload=payload,
-            source=source,
-            priority=priority_enum,
-            metadata=metadata or {},
-            immediate=immediate
-        )
-
-        if immediate and result:
-            response = TriggerEventResponse(
-                success=result.success,
-                event=EventInfo(
-                    name=result.event.name,
-                    id=result.event.id,
-                    source=result.event.source,
-                    timestamp=result.event.timestamp,
-                    priority=result.event.priority.name.lower()
-                ),
-                queued=False,
-                processed=True,
-                routed_to=result.routed_to,
-                handler_name=result.handler_name,
-                error=result.error
-            )
-        else:
-            # Event was queued (not yet processed, success unknown)
-            response = TriggerEventResponse(
-                success=True,  # Queuing succeeded, not event processing
-                queued=True,
-                processed=False,
-                error="Event queued for async processing; success indicates queue operation, not event handling"
-            )
-
-        logger.info(f"Triggered workflow event: {event_name}")
-        return response.model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error triggering workflow event: {e}")
-        return TriggerEventResponse(
-            success=False,
-            error=str(e)
-        ).model_dump_json(indent=2, exclude_none=True)
-
-
-@mcp.tool()
-async def list_workflow_events(ctx: Context) -> str:
-    """List all registered workflow events.
-
-    Returns information about all events that have listeners, routers,
-    or start handlers registered.
-
-    Returns:
-        JSON response with list of registered events
-    """
-    event_bus: EventBus = ctx.request_context.lifespan_context["event_bus"]
-    flow_manager: FlowManager = ctx.request_context.lifespan_context["flow_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Get all registered event names
-        event_names = await event_bus.get_registered_events()
-
-        # Build detailed info for each event
-        events = []
-        for name in sorted(event_names):
-            listeners = await event_bus._registry.get_listeners(name)
-            router = await event_bus._registry.get_router(name)
-            start_handler = await event_bus._registry.get_start_handler(name)
-
-            events.append(WorkflowEventInfo(
-                event_name=name,
-                has_listeners=len(listeners) > 0,
-                has_router=router is not None,
-                is_start_event=start_handler is not None,
-                listener_count=len(listeners)
-            ))
-
-        # Get registered flows
-        flow_names = flow_manager.list_flows()
-
-        response = ListWorkflowEventsResponse(
-            events=events,
-            total_count=len(events),
-            flows_registered=flow_names
-        )
-
-        logger.info(f"Listed {len(events)} workflow events")
-        return response.model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error listing workflow events: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
 
 
 # ============================================================================
 # MEMORY STORE TOOLS
+# See iterm_mcpy/tools/memory.py
 # ============================================================================
-
-# Pattern for safe namespace and key characters
-# Allows alphanumeric, underscore, hyphen, and dot
-_SAFE_MEMORY_PATTERN = re.compile(r'^[a-zA-Z0-9_\-\.]+$')
-
-
-def _validate_namespace(namespace: List[str]) -> None:
-    """Validate namespace parts contain only safe characters.
-
-    Args:
-        namespace: List of namespace parts
-
-    Raises:
-        ValueError: If any part contains invalid characters
-    """
-    if not namespace:
-        return  # Empty namespace is valid (root)
-
-    for part in namespace:
-        if not part:
-            raise ValueError("Namespace parts cannot be empty strings")
-        if not _SAFE_MEMORY_PATTERN.match(part):
-            raise ValueError(
-                f"Invalid namespace part '{part}': only alphanumeric, underscore, hyphen, and dot allowed"
-            )
-
-
-def _validate_key(key: str) -> None:
-    """Validate key contains only safe characters.
-
-    Args:
-        key: The memory key
-
-    Raises:
-        ValueError: If key contains invalid characters
-    """
-    if not key:
-        raise ValueError("Key cannot be empty")
-    if not _SAFE_MEMORY_PATTERN.match(key):
-        raise ValueError(
-            f"Invalid key '{key}': only alphanumeric, underscore, hyphen, and dot allowed"
-        )
-
-
-@mcp.tool()
-async def manage_memory(request: ManageMemoryRequest, ctx: Context) -> str:
-    """Unified memory store operations - consolidates 8 memory tools into one.
-
-    Operations:
-    - store: Save a value (requires namespace, key, value; optional metadata)
-    - retrieve: Get a value (requires namespace, key)
-    - search: Full-text search (requires namespace, query; optional limit)
-    - list_keys: List all keys in namespace (requires namespace)
-    - list_namespaces: List namespaces (optional namespace as prefix filter)
-    - delete: Delete a key (requires namespace, key)
-    - clear: Clear namespace (requires namespace, confirm=True)
-    - stats: Get store statistics (no params required)
-
-    Args:
-        request: ManageMemoryRequest with operation and relevant parameters
-
-    Returns:
-        JSON with operation result
-    """
-    memory_store_instance = ctx.request_context.lifespan_context.get("memory_store")
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    if not memory_store_instance:
-        return ManageMemoryResponse(
-            operation=request.operation,
-            success=False,
-            error="Memory store not initialized"
-        ).model_dump_json(indent=2, exclude_none=True)
-
-    try:
-        op = request.operation
-
-        # STORE operation
-        if op == "store":
-            if not request.namespace:
-                raise ValueError("namespace is required for store operation")
-            if not request.key:
-                raise ValueError("key is required for store operation")
-            if request.value is None:
-                raise ValueError("value is required for store operation")
-
-            _validate_namespace(request.namespace)
-            _validate_key(request.key)
-            ns_tuple = tuple(request.namespace)
-            await memory_store_instance.store(ns_tuple, request.key, request.value, request.metadata)
-            logger.info(f"Stored memory: {'/'.join(request.namespace)}/{request.key}")
-
-            return ManageMemoryResponse(
-                operation=op,
-                success=True,
-                data={
-                    "status": "stored",
-                    "namespace": request.namespace,
-                    "key": request.key,
-                    "metadata": request.metadata or {}
-                }
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # RETRIEVE operation
-        elif op == "retrieve":
-            if not request.namespace:
-                raise ValueError("namespace is required for retrieve operation")
-            if not request.key:
-                raise ValueError("key is required for retrieve operation")
-
-            _validate_namespace(request.namespace)
-            _validate_key(request.key)
-            ns_tuple = tuple(request.namespace)
-            memory = await memory_store_instance.retrieve(ns_tuple, request.key)
-
-            if memory:
-                logger.info(f"Retrieved memory: {'/'.join(request.namespace)}/{request.key}")
-                return ManageMemoryResponse(
-                    operation=op,
-                    success=True,
-                    data={
-                        "found": True,
-                        "key": memory.key,
-                        "value": memory.value,
-                        "timestamp": memory.timestamp.isoformat(),
-                        "metadata": memory.metadata,
-                        "namespace": list(memory.namespace)
-                    }
-                ).model_dump_json(indent=2, exclude_none=True)
-            else:
-                logger.info(f"Memory not found: {'/'.join(request.namespace)}/{request.key}")
-                return ManageMemoryResponse(
-                    operation=op,
-                    success=True,
-                    data={"found": False, "namespace": request.namespace, "key": request.key}
-                ).model_dump_json(indent=2, exclude_none=True)
-
-        # SEARCH operation
-        elif op == "search":
-            if not request.namespace:
-                raise ValueError("namespace is required for search operation")
-            if not request.query:
-                raise ValueError("query is required for search operation")
-
-            _validate_namespace(request.namespace)
-            ns_tuple = tuple(request.namespace)
-            results = await memory_store_instance.search(ns_tuple, request.query, request.limit)
-            logger.info(f"Memory search '{request.query}' in {'/'.join(request.namespace)}: {len(results)} results")
-
-            return ManageMemoryResponse(
-                operation=op,
-                success=True,
-                data={
-                    "query": request.query,
-                    "namespace": request.namespace,
-                    "count": len(results),
-                    "results": [
-                        {
-                            "key": r.memory.key,
-                            "value": r.memory.value,
-                            "score": r.score,
-                            "match_context": r.match_context,
-                            "timestamp": r.memory.timestamp.isoformat(),
-                            "metadata": r.memory.metadata,
-                            "namespace": list(r.memory.namespace)
-                        }
-                        for r in results
-                    ]
-                }
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # LIST_KEYS operation
-        elif op == "list_keys":
-            if not request.namespace:
-                raise ValueError("namespace is required for list_keys operation")
-
-            _validate_namespace(request.namespace)
-            ns_tuple = tuple(request.namespace)
-            keys = await memory_store_instance.list_keys(ns_tuple)
-            logger.info(f"Listed {len(keys)} keys in namespace {'/'.join(request.namespace)}")
-
-            return ManageMemoryResponse(
-                operation=op,
-                success=True,
-                data={"namespace": request.namespace, "count": len(keys), "keys": keys}
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # LIST_NAMESPACES operation
-        elif op == "list_namespaces":
-            if request.namespace:
-                _validate_namespace(request.namespace)
-            prefix_tuple = tuple(request.namespace) if request.namespace else None
-            namespaces = await memory_store_instance.list_namespaces(prefix_tuple)
-            logger.info(f"Listed {len(namespaces)} namespaces")
-
-            return ManageMemoryResponse(
-                operation=op,
-                success=True,
-                data={
-                    "prefix": request.namespace,
-                    "count": len(namespaces),
-                    "namespaces": [list(ns) for ns in namespaces]
-                }
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # DELETE operation
-        elif op == "delete":
-            if not request.namespace:
-                raise ValueError("namespace is required for delete operation")
-            if not request.key:
-                raise ValueError("key is required for delete operation")
-
-            _validate_namespace(request.namespace)
-            _validate_key(request.key)
-            ns_tuple = tuple(request.namespace)
-            deleted = await memory_store_instance.delete(ns_tuple, request.key)
-
-            if deleted:
-                logger.info(f"Deleted memory: {'/'.join(request.namespace)}/{request.key}")
-            else:
-                logger.info(f"Memory not found for deletion: {'/'.join(request.namespace)}/{request.key}")
-
-            return ManageMemoryResponse(
-                operation=op,
-                success=True,
-                data={
-                    "deleted": deleted,
-                    "namespace": request.namespace,
-                    "key": request.key,
-                    "message": None if deleted else "Memory not found"
-                }
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # CLEAR operation
-        elif op == "clear":
-            if not request.namespace:
-                raise ValueError("namespace is required for clear operation")
-
-            _validate_namespace(request.namespace)
-
-            if not request.confirm:
-                return ManageMemoryResponse(
-                    operation=op,
-                    success=False,
-                    error="Confirmation required. Set confirm=True to clear namespace. This permanently deletes all memories.",
-                    data={"namespace": request.namespace}
-                ).model_dump_json(indent=2, exclude_none=True)
-
-            ns_tuple = tuple(request.namespace)
-            count = await memory_store_instance.clear_namespace(ns_tuple)
-            logger.info(f"Cleared namespace {'/'.join(request.namespace)}: {count} memories deleted")
-
-            return ManageMemoryResponse(
-                operation=op,
-                success=True,
-                data={"cleared": True, "namespace": request.namespace, "deleted_count": count}
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # STATS operation
-        elif op == "stats":
-            stats = await memory_store_instance.get_stats()
-            logger.info("Retrieved memory store stats")
-
-            return ManageMemoryResponse(
-                operation=op,
-                success=True,
-                data=stats
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        else:
-            return ManageMemoryResponse(
-                operation=op,
-                success=False,
-                error=f"Unknown operation: {op}"
-            ).model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error in manage_memory ({request.operation}): {e}")
-        return ManageMemoryResponse(
-            operation=request.operation,
-            success=False,
-            error=str(e)
-        ).model_dump_json(indent=2, exclude_none=True)
-
-
-@mcp.tool()
-async def get_workflow_event_history(
-    ctx: Context,
-    event_name: Optional[str] = None,
-    limit: int = 100,
-    success_only: bool = False
-) -> str:
-    """Get workflow event history.
-
-    Args:
-        event_name: Filter by event name (optional)
-        limit: Max entries to return (default: 100, max: 1000)
-        success_only: Only return successfully processed events
-
-    Returns:
-        JSON response with event history entries
-    """
-    event_bus: EventBus = ctx.request_context.lifespan_context["event_bus"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Cap limit
-        limit = min(limit, 1000)
-
-        # Get history
-        history = await event_bus.get_history(
-            event_name=event_name,
-            limit=limit,
-            success_only=success_only
-        )
-
-        # Convert to response format
-        entries = [
-            EventHistoryEntry(
-                event_name=r.event.name,
-                event_id=r.event.id,
-                source=r.event.source,
-                timestamp=r.event.timestamp,
-                success=r.success,
-                handler_name=r.handler_name,
-                routed_to=r.routed_to,
-                duration_ms=r.duration_ms,
-                error=r.error
-            )
-            for r in history
-        ]
-
-        response = GetEventHistoryResponse(
-            entries=entries,
-            total_count=len(entries)
-        )
-
-        logger.info(f"Retrieved {len(entries)} event history entries")
-        return response.model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error getting event history: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def subscribe_to_output_pattern(
-    ctx: Context,
-    pattern: str,
-    event_name: Optional[str] = None
-) -> str:
-    """Subscribe to terminal output matching a pattern.
-
-    When terminal output matches the pattern, the specified event will be
-    triggered with the matched text as payload.
-
-    Args:
-        pattern: Regex pattern to match against terminal output
-        event_name: Event to trigger on pattern match (optional)
-
-    Returns:
-        JSON response with subscription ID
-    """
-    event_bus: EventBus = ctx.request_context.lifespan_context["event_bus"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Validate pattern
-        re.compile(pattern)
-
-        # Create callback
-        async def on_match(text: str, match: Any) -> None:
-            logger.debug(f"Pattern matched: {pattern} -> {match}")
-
-        # Subscribe
-        subscription_id = await event_bus.subscribe_to_pattern(
-            pattern=pattern,
-            callback=on_match,
-            event_name=event_name
-        )
-
-        response = PatternSubscriptionResponse(
-            subscription_id=subscription_id,
-            pattern=pattern,
-            event_name=event_name
-        )
-
-        logger.info(f"Created pattern subscription: {pattern}")
-        return response.model_dump_json(indent=2, exclude_none=True)
-
-    except re.error as e:
-        logger.error(f"Invalid regex pattern: {e}")
-        return json.dumps({"error": f"Invalid regex pattern: {e}"}, indent=2)
-    except Exception as e:
-        logger.error(f"Error creating pattern subscription: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
 
 
 # ============================================================================
 # AGENT HOOKS MANAGEMENT
+# See iterm_mcpy/tools/agent_hooks.py
 # ============================================================================
-
-@mcp.tool()
-async def manage_agent_hooks(request: ManageAgentHooksRequest, ctx: Context) -> str:
-    """Unified agent hooks management.
-
-    Operations:
-    - get_config: Get current global hooks configuration
-    - update_config: Update global hooks configuration
-    - get_repo_config: Get hooks config for a specific repo (.iterm/hooks.json)
-    - trigger_path_change: Manually trigger path change hook for a session
-    - get_stats: Get hook manager statistics
-    - set_variable: Set an iTerm user variable to enable/disable hooks per session
-    - get_variable: Get an iTerm user variable value
-
-    Args:
-        request: ManageAgentHooksRequest with operation and parameters
-
-    Returns:
-        JSON with operation result
-    """
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        from core.agent_hooks import get_agent_hook_manager, GlobalHooksConfig
-
-        hook_manager = get_agent_hook_manager()
-        op = request.operation
-
-        # GET_CONFIG operation
-        if op == "get_config":
-            config = hook_manager.config
-            return ManageAgentHooksResponse(
-                operation=op,
-                success=True,
-                data=config.model_dump()
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # UPDATE_CONFIG operation
-        elif op == "update_config":
-            config_updates = {}
-            if request.enabled is not None:
-                config_updates["enabled"] = request.enabled
-            if request.auto_team_assignment is not None:
-                config_updates["auto_team_assignment"] = request.auto_team_assignment
-            if request.fallback_team_from_repo is not None:
-                config_updates["fallback_team_from_repo"] = request.fallback_team_from_repo
-            if request.pass_session_id_default is not None:
-                config_updates["pass_session_id_default"] = request.pass_session_id_default
-
-            if config_updates:
-                # Update config
-                current_data = hook_manager.config.model_dump()
-                current_data.update(config_updates)
-                hook_manager.config = GlobalHooksConfig(**current_data)
-                hook_manager.save_global_config()
-                logger.info(f"Updated agent hooks config: {config_updates}")
-
-            return ManageAgentHooksResponse(
-                operation=op,
-                success=True,
-                data={
-                    "updated": config_updates,
-                    "config": hook_manager.config.model_dump()
-                }
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # GET_REPO_CONFIG operation
-        elif op == "get_repo_config":
-            if not request.repo_path:
-                return ManageAgentHooksResponse(
-                    operation=op,
-                    success=False,
-                    error="repo_path is required for get_repo_config operation"
-                ).model_dump_json(indent=2, exclude_none=True)
-
-            repo_config = hook_manager.load_repo_config(request.repo_path)
-            return ManageAgentHooksResponse(
-                operation=op,
-                success=True,
-                data={
-                    "repo_path": request.repo_path,
-                    "config": repo_config.model_dump() if repo_config else None,
-                    "config_file": str(Path(request.repo_path) / hook_manager.config.repo_config_filename)
-                }
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # TRIGGER_PATH_CHANGE operation
-        elif op == "trigger_path_change":
-            if not request.session_id:
-                return ManageAgentHooksResponse(
-                    operation=op,
-                    success=False,
-                    error="session_id is required for trigger_path_change operation"
-                ).model_dump_json(indent=2, exclude_none=True)
-            if not request.new_path:
-                return ManageAgentHooksResponse(
-                    operation=op,
-                    success=False,
-                    error="new_path is required for trigger_path_change operation"
-                ).model_dump_json(indent=2, exclude_none=True)
-
-            result = await hook_manager.on_path_changed(
-                request.session_id,
-                request.new_path,
-                request.agent_name
-            )
-            logger.info(f"Triggered path change for {request.session_id}: {result.actions_taken}")
-
-            return ManageAgentHooksResponse(
-                operation=op,
-                success=True,
-                data=result.to_dict()
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # GET_STATS operation
-        elif op == "get_stats":
-            stats = hook_manager.get_stats()
-            return ManageAgentHooksResponse(
-                operation=op,
-                success=True,
-                data=stats
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # SET_VARIABLE operation - Set iTerm user variable to enable/disable hooks
-        elif op == "set_variable":
-            if not request.session_id:
-                return ManageAgentHooksResponse(
-                    operation=op,
-                    success=False,
-                    error="session_id is required for set_variable operation"
-                ).model_dump_json(indent=2, exclude_none=True)
-            if not request.variable_name:
-                return ManageAgentHooksResponse(
-                    operation=op,
-                    success=False,
-                    error="variable_name is required for set_variable operation"
-                ).model_dump_json(indent=2, exclude_none=True)
-
-            from core.iterm_path_monitor import set_user_variable
-            connection = ctx.request_context.lifespan_context["connection"]
-            success = await set_user_variable(
-                connection,
-                request.session_id,
-                request.variable_name,
-                request.variable_value or ""
-            )
-            logger.info(f"Set variable {request.variable_name}={request.variable_value} on session {request.session_id}: success={success}")
-
-            return ManageAgentHooksResponse(
-                operation=op,
-                success=success,
-                data={
-                    "session_id": request.session_id,
-                    "variable_name": request.variable_name,
-                    "variable_value": request.variable_value
-                },
-                error=None if success else "Failed to set variable (session may not exist or iTerm not connected)"
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # GET_VARIABLE operation - Get iTerm user variable
-        elif op == "get_variable":
-            if not request.session_id:
-                return ManageAgentHooksResponse(
-                    operation=op,
-                    success=False,
-                    error="session_id is required for get_variable operation"
-                ).model_dump_json(indent=2, exclude_none=True)
-            if not request.variable_name:
-                return ManageAgentHooksResponse(
-                    operation=op,
-                    success=False,
-                    error="variable_name is required for get_variable operation"
-                ).model_dump_json(indent=2, exclude_none=True)
-
-            from core.iterm_path_monitor import get_user_variable
-            connection = ctx.request_context.lifespan_context["connection"]
-            value = await get_user_variable(connection, request.session_id, request.variable_name)
-            logger.info(f"Got variable {request.variable_name} from session {request.session_id}: {value}")
-
-            return ManageAgentHooksResponse(
-                operation=op,
-                success=True,
-                data={
-                    "session_id": request.session_id,
-                    "variable_name": request.variable_name,
-                    "variable_value": value
-                }
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        else:
-            return ManageAgentHooksResponse(
-                operation=op,
-                success=False,
-                error=f"Unknown operation: {op}"
-            ).model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error in manage_agent_hooks ({request.operation}): {e}")
-        return ManageAgentHooksResponse(
-            operation=request.operation,
-            success=False,
-            error=str(e)
-        ).model_dump_json(indent=2, exclude_none=True)
 
 
 @mcp.resource("memory://stats")
