+++ iterm_mcpy/tools/__init__.py
@@ -9,6 +9,8 @@ def register_all(mcp):
     """Register all tool modules with the FastMCP instance.
 
     Called from fastmcp_server.py after mcp creation and before run().
+    Tool modules are added incrementally; this file is updated as each
+    batch of modules is extracted from fastmcp_server.py.
     """
     from . import (
         memory,
@@ -16,26 +18,10 @@ def register_all(mcp):
         telemetry,
         workflows,
         services,
-        roles,
-        notifications,
-        wait,
-        feedback,
-        managers,
-        agents,
-        control,
-        monitoring,
-        modifications,
-        orchestration,
-        commands,
-        sessions,
     )
 
     _MODULES = [
-        memory, agent_hooks, telemetry, workflows,
-        services, roles, notifications, wait,
-        feedback, managers, agents, control,
-        monitoring, modifications, orchestration,
-        commands, sessions,
+        memory, agent_hooks, telemetry, workflows, services,
     ]
 
     for mod in _MODULES:
